Code cleanup suggested by Eclipse.
--HG--
branch : r1.1.0

+++ org.jenetics/src/main/java/org/jenetics/stat/Variance.java
@@ -63,7 +63,7 @@ public class Variance<N extends Number> extends Mean<N> {
 		if (_samples == 1) {
 			variance = _m2;
 		} else if (_samples > 1) {
-			variance = _m2/(double)(_samples - 1);
+			variance = _m2/(_samples - 1);
 		}
 
 		return variance;
@@ -82,7 +82,7 @@ public class Variance<N extends Number> extends Mean<N> {
 		final double data = value.doubleValue();
 		final double delta = data - _mean;
 
-		_mean += delta/(double)(++_samples);
+		_mean += delta/(++_samples);
 		_m2 += delta*(data - _mean);
 	}
